+++ asteroid-test/src/main/groovy/asteroid/local/samples/ExecuteMethodSafelyImpl.groovy
@@ -15,7 +15,7 @@ class ExecuteMethodSafelyImpl extends AbstractLocalTransformation<ExecuteMethodS
 
     @Override
     void doVisit(AnnotationNode annotation, MethodNode methodNode) {
-        List<FieldNode> fields = A.UTIL.CLASS.getInstancePropertyFields(methodNode.declaringClass)
+        List<FieldNode> fields = A.UTIL.NODE.getInstancePropertyFields(methodNode.declaringClass)
         String valueFieldNode = fields.find()?.name
 
         methodNode.code = A.STMT.returnS(A.EXPR.safeCallX(A.EXPR.varX(valueFieldNode), 'toString'))