+++ asteroid-test/src/main/groovy/asteroid/local/samples/ToMapImpl.groovy
@@ -19,16 +19,16 @@ class ToMapImpl extends AbstractLocalTransformation<ToMap, ClassNode> {
 
     @Override
     void doVisit(AnnotationNode annotation, ClassNode node) {
-        List<FieldNode> instanceFields   = A.UTIL.CLASS.getInstancePropertyFields(node)
+        List<FieldNode> instanceFields   = A.UTIL.NODE.getInstancePropertyFields(node)
         List<MapEntryExpression> entries = instanceFields.collect(this.&fieldToMapEntry)
-        MapExpression mapExpression      = new MapExpression(entries)
+        MapExpression mapExpression      = A.EXPR.mapX(entries)
         MethodNode methodNode            = getToMapMethod(mapExpression)
 
-        A.UTIL.CLASS.addMethod(node, methodNode)
+        A.UTIL.NODE.addMethod(node, methodNode)
     }
 
     private MapEntryExpression fieldToMapEntry(final FieldNode field) {
-        return new MapEntryExpression(A.EXPR.constX(field.name), A.EXPR.varX(field.name))
+        return A.EXPR.mapEntryX(A.EXPR.constX(field.name), A.EXPR.varX(field.name))
     }
 
     private MethodNode getToMapMethod(final MapExpression map) {